Force list-cache generation if List Manager enabled lists have changed.
This addresses a problem in that if a user is update-limited, then they disable a list in their List Manager, then re-enable that same list, then close the List Manager . . . they would not get that list loaded.  We'll now load the already-downloaded copy of that list, even if we can't download a new copy.